requirements.txt
selenium==4.21.0
bs4==0.0.2
psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
SQLAlchemy==2.0.30
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
pytest==8.2.1
pytest-flask==1.3.0
pytest-mock==3.14.0
